+++ data-raw/create_constants.R
@@ -113,6 +113,7 @@ psut_cols <- list(resources = "R",
                   V = "V", 
                   final_demand = "Y",
                   Y = "Y", 
+                  epsilon = "Epsilon",
                   s_units = "S_units", 
                   matvals = "matvals")
 usethis::use_data(psut_cols, overwrite = TRUE)