+++ tests/testthat/test-psut.R
@@ -45,6 +45,46 @@ test_that("add_psut_matnames works as expected", {
   # Transfers
   expect_true(With_matnames %>% dplyr::filter(startsWith(Flow, "Transfers") & E.dot < 0) %>% magrittr::extract2("matnames") %>% magrittr::equals("U_feed") %>% all())
   expect_true(With_matnames %>% dplyr::filter(startsWith(Flow, "Transfers") & E.dot > 0) %>% magrittr::extract2("matnames") %>% magrittr::equals("V") %>% all())
+  
+  With_matnames <- load_tidy_iea_df() %>% 
+    specify_all() %>% 
+    add_psut_matnames() %>% 
+    tibble::add_row(
+      Country = c("GHA", "GHA"),
+      Method = c("PCM", "PCM"),
+      Energy.type = c("E", "E"),
+      Last.stage = c("Final", "Final"),
+      Year = c(1971, 1971),
+      Ledger.side = c("Consumption", "Supply"),
+      Flow.aggregation.point = c("Industry", "Total primary energy supply"),
+      Flow = c("Non-ferrous metals", "Stock changes [of Crude oil]"),
+      Product = c("Electricity", "Crude oil"),
+      Unit = c("ktoe", "ktoe"),
+      E.dot = c(100, -100),
+      matnames = psut_cols$epsilon
+    ) %>% 
+    add_psut_matnames()
+  
+  expect_equal(With_matnames %>% dplyr::filter(matnames == "Epsilon") %>% nrow(), 2)
+  
+  With_matnames_2 <- load_tidy_iea_df() %>% 
+    specify_all() %>% 
+    add_psut_matnames() %>% 
+    tibble::add_row(
+      Country = c("GHA", "GHA"),
+      Method = c("PCM", "PCM"),
+      Energy.type = c("E", "E"),
+      Last.stage = c("Final", "Final"),
+      Year = c(1971, 1971),
+      Ledger.side = c("Consumption", "Supply"),
+      Flow.aggregation.point = c("Industry", "Total primary energy supply"),
+      Flow = c("Non-ferrous metals", "Stock changes [of Crude oil]"),
+      Product = c("Electricity", "Crude oil"),
+      Unit = c("ktoe", "ktoe"),
+      E.dot = c(100, -100),
+      matnames = psut_cols$epsilon)
+  
+  expect_true(all(With_matnames == With_matnames_2))
 })
 
 
@@ -52,6 +92,20 @@ test_that("add_row_col_meta works as expected", {
   With_meta <- load_tidy_iea_df() %>% 
     specify_all() %>% 
     add_psut_matnames() %>% 
+    tibble::add_row(
+      Country = c("GHA", "GHA"),
+      Method = c("PCM", "PCM"),
+      Energy.type = c("E", "E"),
+      Last.stage = c("Final", "Final"),
+      Year = c(1971, 1971),
+      Ledger.side = c("Consumption", "Supply"),
+      Flow.aggregation.point = c("Industry", "Total primary energy supply"),
+      Flow = c("Non-ferrous metals", "Stock changes [of Crude oil]"),
+      Product = c("Electricity", "Crude oil"),
+      Unit = c("ktoe", "ktoe"),
+      E.dot = c(100, -100),
+      matnames = psut_cols$epsilon
+    ) %>% 
     add_row_col_meta()
   # Ensure that every row is filled in the new columns.
   expect_false(any(is.na(With_meta$rownames)))
@@ -63,6 +117,8 @@ test_that("add_row_col_meta works as expected", {
   expect_true(With_meta %>% dplyr::filter(matnames == "R" | matnames == "V") %>% magrittr::extract2("coltypes") %>% magrittr::equals("Product") %>% all())
   expect_true(With_meta %>% dplyr::filter(startsWith(matnames, "U") | matnames == "Y") %>% magrittr::extract2("rowtypes") %>% magrittr::equals("Product") %>% all())
   expect_true(With_meta %>% dplyr::filter(startsWith(matnames, "U") | matnames == "Y") %>% magrittr::extract2("coltypes") %>% magrittr::equals("Industry") %>% all())
+  expect_true(With_meta %>% dplyr::filter(startsWith(matnames, "Epsilon")) %>% magrittr::extract2("rowtypes") %>% magrittr::equals("Product") %>% all())
+  expect_true(With_meta %>% dplyr::filter(startsWith(matnames, "Epsilon")) %>% magrittr::extract2("coltypes") %>% magrittr::equals("Industry") %>% all())
   # Ensure that row and column identifiers are correct
   # Rows of R and V are the Flow names
   magrittr::equals(With_meta %>% dplyr::filter(matnames == "R" | matnames == "V") %>% magrittr::extract2("rownames"), 
@@ -79,11 +135,118 @@ test_that("add_row_col_meta works as expected", {
                    With_meta %>% dplyr::filter(matnames == "U" | matnames == "Y") %>% magrittr::extract2("Product")) %>% 
     all() %>% 
     expect_true()
+  # Rows of Epsilon are the Product names
+  magrittr::equals(With_meta %>% dplyr::filter(matnames == "Epsilon") %>% magrittr::extract2("rownames"), 
+                   With_meta %>% dplyr::filter(matnames == "Epsilon") %>% magrittr::extract2("Product")) %>% 
+    all() %>% 
+    expect_true()
+  
   # Columns of U and Y are the Flow names
   magrittr::equals(With_meta %>% dplyr::filter(matnames == "U" | matnames == "Y") %>% magrittr::extract2("colnames"), 
                    With_meta %>% dplyr::filter(matnames == "U" | matnames == "Y") %>% magrittr::extract2("Flow")) %>% 
     all() %>% 
     expect_true()
+  
+  magrittr::equals(With_meta %>% dplyr::filter(matnames == "Epsilon") %>% magrittr::extract2("colnames"), 
+                   With_meta %>% dplyr::filter(matnames == "Epsilon") %>% magrittr::extract2("Flow")) %>% 
+    all() %>% 
+    expect_true()
+  
+  # If columns are already there, add_row_col_meta() shouldn't do anything.
+  With_meta <- load_tidy_iea_df() %>% 
+    specify_all() %>% 
+    add_psut_matnames() %>% 
+    tibble::add_row(
+      Country = c("GHA", "GHA"),
+      Method = c("PCM", "PCM"),
+      Energy.type = c("E", "E"),
+      Last.stage = c("Final", "Final"),
+      Year = c(1971, 1971),
+      Ledger.side = c("Consumption", "Supply"),
+      Flow.aggregation.point = c("Industry", "Total primary energy supply"),
+      Flow = c("Non-ferrous metals", "Stock changes [of Crude oil]"),
+      Product = c("Electricity", "Crude oil"),
+      Unit = c("ktoe", "ktoe"),
+      E.dot = c(100, -100),
+      matnames = psut_cols$epsilon
+    ) %>% 
+    tibble::add_column(
+      rownames = "a very odd name",
+      colnames = "a WEIRD name",
+      rowtypes = "a damn weird row type",
+      coltypes = "a bloody strange col type"
+    )
+  
+  With_meta_2 <- load_tidy_iea_df() %>% 
+    specify_all() %>% 
+    add_psut_matnames() %>% 
+    tibble::add_row(
+      Country = c("GHA", "GHA"),
+      Method = c("PCM", "PCM"),
+      Energy.type = c("E", "E"),
+      Last.stage = c("Final", "Final"),
+      Year = c(1971, 1971),
+      Ledger.side = c("Consumption", "Supply"),
+      Flow.aggregation.point = c("Industry", "Total primary energy supply"),
+      Flow = c("Non-ferrous metals", "Stock changes [of Crude oil]"),
+      Product = c("Electricity", "Crude oil"),
+      Unit = c("ktoe", "ktoe"),
+      E.dot = c(100, -100),
+      matnames = psut_cols$epsilon
+    ) %>% 
+    tibble::add_column(
+      rownames = "a very odd name",
+      colnames = "a WEIRD name",
+      rowtypes = "a damn weird row type",
+      coltypes = "a bloody strange col type"
+    ) %>% 
+    add_row_col_meta() %>% 
+    magrittr::equals(With_meta) %>% 
+    all() %>% 
+    expect_true()
+  
+  # Now, expect error if only one of the 4 expected columns is present.
+  With_meta <- load_tidy_iea_df() %>% 
+    specify_all() %>% 
+    add_psut_matnames() %>% 
+    dplyr::mutate(
+      "{IEATools::mat_meta_cols$rownames}" := "a very odd rowname"
+    ) %>% 
+    add_row_col_meta() %>% 
+    expect_error(regexp = "'rownames' is")
+  
+  
+  # Verify that any added column triggers the error. 
+  # This time do "colnames"
+  load_tidy_iea_df() %>% 
+    specify_all() %>% 
+    add_psut_matnames() %>% 
+    dplyr::mutate(
+      "{IEATools::mat_meta_cols$colnames}" := "a bad colname"
+    ) %>% 
+    add_row_col_meta() %>% 
+    testthat::expect_error(regexp = "'colnames' is ")
+
+  # rowtypes  
+  load_tidy_iea_df() %>% 
+    specify_all() %>% 
+    add_psut_matnames() %>% 
+    dplyr::mutate(
+      "{IEATools::mat_meta_cols$rowtypes}" := "a bad rowtype"
+    ) %>% 
+    add_row_col_meta() %>% 
+    testthat::expect_error(regexp = "'rowtypes' is ")
+
+  # coltypes  
+  load_tidy_iea_df() %>% 
+    specify_all() %>% 
+    add_psut_matnames() %>% 
+    dplyr::mutate(
+      "{IEATools::mat_meta_cols$coltypes}" := "a bad coltype"
+    ) %>% 
+    add_row_col_meta() %>% 
+    testthat::expect_error(regexp = "'coltypes' is ")
+
 })
 
 
@@ -91,11 +254,26 @@ test_that("collapse_to_psut works expected", {
   With_mats <- load_tidy_iea_df() %>% 
     specify_all() %>% 
     add_psut_matnames() %>% 
+    tibble::add_row(
+      Country = c("GHA", "GHA"),
+      Method = c("PCM", "PCM"),
+      Energy.type = c("E", "E"),
+      Last.stage = c("Final", "Final"),
+      Year = c(1971, 1971),
+      Ledger.side = c("Consumption", "Supply"),
+      Flow.aggregation.point = c("Industry", "Total primary energy supply"),
+      Flow = c("Non-ferrous metals", "Stock changes [of Crude oil]"),
+      Product = c("Electricity", "Crude oil"),
+      Unit = c("ktoe", "ktoe"),
+      E.dot = c(100, -100),
+      matnames = psut_cols$epsilon
+    ) %>% 
     add_row_col_meta() %>% 
     collapse_to_tidy_psut()
-  expect_equal(nrow(With_mats), 20)
-  # Ensure that all values in the matrices are positive.
+  expect_equal(nrow(With_mats), 21)
+  # Ensure that all values in the matrices (excluding Epsilon) are positive.
   With_mats %>%
+    dplyr::filter(matnames != "Epsilon") %>% 
     dplyr::mutate(
       gezero = matsbyname::compare_byname(matvals, ">=", 0) %>% matsbyname::all_byname()
     ) %>% 
@@ -157,7 +335,7 @@ test_that("prep_psut() works as expected with empty .tidy_iea_df", {
                      IEATools::iea_cols$energy_type,
                      IEATools::iea_cols$last_stage,
                      IEATools::iea_cols$year,
-                     "R", "U_EIOU", "U_feed", "V", "Y", "S_units")) {
+                     "R", "U_EIOU", "U_feed", "V", "Y", "S_units", "Epsilon")) {
     expect_true(col_name %in% cn)
   }
   expect_equal(nrow(zero_psut), 0)
@@ -178,3 +356,45 @@ test_that("prep_psut() correctly makes columns of U and r_EIOU matrices", {
   expect_equal(psut_with_test_cols[["U_test"]], psut_with_test_cols[["U"]])
   expect_equal(psut_with_test_cols[["r_EIOU_test"]], psut_with_test_cols[["r_EIOU"]])
 })
+
+
+test_that("prep_psut() correctly works with Epsilon flows", {
+  
+  PSUT_flows_with_Epsilon <- load_tidy_iea_df() %>% 
+    specify_all() %>% 
+    add_psut_matnames() %>% 
+    tibble::add_row(
+      Country = c("GHA", "GHA"),
+      Method = c("PCM", "PCM"),
+      Energy.type = c("E", "E"),
+      Last.stage = c("Final", "Final"),
+      Year = c(1971, 1971),
+      Ledger.side = c("Consumption", "Supply"),
+      Flow.aggregation.point = c("Industry", "Total primary energy supply"),
+      Flow = c("Non-ferrous metals", "Stock changes [of Crude oil]"),
+      Product = c("Electricity", "Crude oil"),
+      Unit = c("ktoe", "ktoe"),
+      E.dot = c(100, -100),
+      matnames = c(psut_cols$epsilon, psut_cols$epsilon)
+    ) %>%
+    prep_psut()
+
+  expect_true(
+    all(c("R", "V", "U_feed", "U_EIOU", "Y", "S_units", "Epsilon") %in% colnames(PSUT_flows_with_Epsilon))
+  )
+  
+  epsilon_expected_value = matrix(nrow = 2, ncol = 2,
+                                  c(0, 100, -100, 0))
+  
+  expect_true(
+    all(epsilon_expected_value == (PSUT_flows_with_Epsilon %>%
+                                     dplyr::select(Epsilon) %>% 
+                                     dplyr::pull() %>% 
+                                     dplyr::first())
+  ))
+})
+
+
+
+
+